+++ src/main/java/br/com/sgp/adapters/inbound/request/ProdutoRequest.java
@@ -3,11 +3,13 @@ package br.com.sgp.adapters.inbound.request;
 import lombok.Getter;
 import lombok.Setter;
 
+import java.math.BigDecimal;
+
 @Getter
 @Setter
 public class ProdutoRequest {
 
-    protected Double valor;
+    protected BigDecimal valor;
     protected Boolean entregue;
     protected Boolean chegou;
