+++ src/main/java/br/com/sgp/application/core/domain/Pedido.java
@@ -3,8 +3,8 @@ package br.com.sgp.application.core.domain;
 import lombok.Getter;
 import lombok.Setter;
 
+import java.math.BigDecimal;
 import java.time.OffsetDateTime;
-import java.util.Date;
 import java.util.List;
 
 @Getter
@@ -13,10 +13,10 @@ public class Pedido {
 
     private Long id;
     private OffsetDateTime data;
-    private Double valor;
+    private BigDecimal valor;
     private StatusPedido situacao;
     private StatusPagamento statusPagamento;
-    private Double valorPago;
+    private BigDecimal valorPago;
     private Temporada temporada;
     private Aluno aluno;
     private List<Produto> produtos;
